Adding in error checking

+++ bin/main.js
@@ -11,7 +11,6 @@ const catalogCollection=process.env.CATALOG_COLLECTION;
 const stockCollection=process.env.STOCK_COLLECTION;
 
 
-
 //Chris's Work List
 
 //Then
@@ -91,6 +90,22 @@ async function main(){
 //Create a new item in the catalog
 async function createItem(newCatalog){
 
+    if (typeof newCatalog.name == "undefined"){
+        return "Error: Item had no name."
+    }
+    else  if (typeof newCatalog.description == "undefined"){
+        return "Error: Item had no description."
+    }
+    else  if (typeof newCatalog.category == "undefined"){
+        return "Error: Item had no category."
+    }
+    else  if (typeof newCatalog.price == "undefined"){
+        return "Error: Item had no price."
+    }
+    else  if (typeof newCatalog.stock == "undefined"){
+        return "Error: Item had no stock."
+    }
+
     id=await generateSKU()
 
     const client = new MongoClient(uri);
@@ -114,7 +129,7 @@ async function createItem(newCatalog){
                 name: newName,
                 stock: newStock.stock
             })
-            return
+            return "Item already existed, so new batch created instead"
         }
 
         const result=await client.db(nameOfDatabase).collection(catalogCollection).insertOne(newCatalog)
@@ -128,7 +143,7 @@ async function createItem(newCatalog){
             })
         }
 
-        console.log(`New Listing made with id of: ${result.insertedId}`)
+        return `New Listing made with id of: ${result.insertedId}`
     } catch (e) {
         console.error(e);
     } finally {
@@ -143,9 +158,14 @@ async function removeItem(removeCatalog){
         // Connect to the MongoDB cluster
         await client.connect();
         deleting=await client.db(nameOfDatabase).collection(catalogCollection).findOne({name: removeCatalog})
+        if (deleting == null){
+            return "Error: Name does not exist"
+        }
+        itemName=deleting.name
         await client.db(nameOfDatabase).collection(catalogCollection).deleteOne({name: removeCatalog})
         id=deleting._id
         await client.db(nameOfDatabase).collection(stockCollection).deleteOne({_id: id})
+        return `Deleted item: ${itemName}`
     } catch (e) {
         console.error(e);
     } finally {
@@ -169,6 +189,12 @@ async function getStock(client, id){
 
 //Add a batch given the stock of batch, name of item
 async function createBatch(newBatch){
+    if (typeof newBatch.name == "undefined"){
+        return "Error: Need an item name."
+    }
+    else  if (typeof newBatch.stock == "undefined"){
+        return "Error: Batch had no stock."
+    }
     const client = new MongoClient(uri);
     try {
         // Connect to the MongoDB cluster
@@ -176,6 +202,9 @@ async function createBatch(newBatch){
         itemName=newBatch.name
         newStock=newBatch.stock
         catalogItem= await client.db(nameOfDatabase).collection(catalogCollection).findOne({name: itemName})
+        if (catalogItem == null){
+            return "Error: Item name does not exist"
+        }
         SKU=catalogItem._id
         stockItem=await client.db(nameOfDatabase).collection(stockCollection).findOne({_id: SKU})
         await client.db(nameOfDatabase).collection(stockCollection).updateOne({_id: SKU},{$set: {stock: newStock+stockItem.stock}})
@@ -192,6 +221,8 @@ async function createBatch(newBatch){
         await client.db(nameOfDatabase).collection(stockCollection).updateOne({_id: SKU},{$set: {batch: batch}})
 
         await client.db(nameOfDatabase).collection(stockCollection).updateOne({_id: SKU},{$rename: {batch: "batch".concat(id)}})
+
+        return `New batch created with id: ${id}`
     } catch (e) {
         console.error(e);
     } finally {
@@ -201,6 +232,9 @@ async function createBatch(newBatch){
 
 //Remove a batch given the batch id
 async function removeBatch(batchId){
+    if (typeof batchId._id == "undefined"){
+        return "Error: Need a batch id."
+    }
     const client = new MongoClient(uri);
     try {
         // Connect to the MongoDB cluster
@@ -211,12 +245,24 @@ async function removeBatch(batchId){
         batch="batch".concat(String(id))
 
         item=await client.db(nameOfDatabase).collection(stockCollection).findOne({[batch]: {$exists: true}})
+        if (item == null){
+            return "Error: Batch ID does not exist"
+        }
 
         //Update stock value and batch count value
         await client.db(nameOfDatabase).collection(stockCollection).updateOne({[batch]: {$exists: true}}, {$set: {stock: item.stock-item[batch].stock}})
         await client.db(nameOfDatabase).collection(stockCollection).updateOne({[batch]: {$exists: true}}, {$set: {batchCount: item.batchCount-1}})
 
         await client.db(nameOfDatabase).collection(stockCollection).updateOne({[batch]: {$exists: true}}, {$unset: {[batch]: ""}})
+
+        if (item.batchCount==1){
+            SKU=item._id
+            remove=await client.db(nameOfDatabase).collection(catalogCollection).findOne({_id: SKU})
+            await removeItem(remove.name)
+            return "Item was removed due to final batch being removed"
+        }
+
+        return `${batch} was removed.`
     } catch (e) {
         console.error(e);
     } finally {
@@ -226,6 +272,12 @@ async function removeBatch(batchId){
 
 //Take some stock away or increase stock in anyt given batch given batch id, and stock change
 async function batchStock(editBatch){
+    if (typeof editBatch._id == "undefined"){
+        return "Error: Need a batch id."
+    }
+    else  if (typeof editBatch.stock == "undefined"){
+        return "Error: Need stock alteration."
+    }
     const client = new MongoClient(uri);
     try {
         // Connect to the MongoDB cluster
@@ -237,12 +289,33 @@ async function batchStock(editBatch){
         batch="batch".concat(String(id))
 
         item=await client.db(nameOfDatabase).collection(stockCollection).findOne({[batch]: {$exists: true}})
+        if (item == null){
+            return "Error: Batch ID does not exist"
+        }
+
+        batchStock=item[batch].stock
+        if (batchStock<=stockChange){
+            await removeBatch({_id: item[batch]._id})
+            if (batchStock==item.stock){
+                SKU=item._id
+                remove=await client.db(nameOfDatabase).collection(catalogCollection).findOne({_id: SKU})
+                await removeItem(remove.name)
+                return "Item was removed due to all stock of final batch being depleted"
+            }
+
+            if (batchStock<stockChange){
+                return `Warning ${batch} did not have ${stockChange} stock remaining. It only had ${batchStock}.`
+            }
+            return "The batch was removed due to no stock remaining."
+        }
 
         await client.db(nameOfDatabase).collection(stockCollection).updateOne({[batch]: {$exists: true}}, {$set: {[batch]: {
             _id: id,
             stock: item[batch].stock-stockChange
         }}})
         await client.db(nameOfDatabase).collection(stockCollection).updateOne({[batch]: {$exists: true}}, {$set: {stock: item.stock-stockChange}})
+
+        return `${batch} stock was altered.`
     } catch (e) {
         console.error(e);
     } finally {
